#17 Move and remove some test cases for LineDataset from utils
